Rename file name(decoder->top_decoder) and remove file gitkeep.
