+++ nn_modules/decoder/top_decoder_test.py
@@ -1,6 +1,6 @@
 import unittest
 import torch
-from decoder import TopDecoder
+from top_decoder import TopDecoder
 
 
 class TestTopDecoder(unittest.TestCase):
